Update package model structure.
