+++ src/main/java/model/BaseModel.java
@@ -1,4 +1,4 @@
-package model.models;
+package model;
 
 import java.util.Objects;
