+++ src/main/java/model/User.java
@@ -1,4 +1,4 @@
-package model.models;
+package model;
 
 import java.time.LocalDate;
 import java.util.Objects;
